Removes the synchronous behavior of zip.
Previous implementation would `await` one `anext` at a time, preventing any concurrency from occurring among the iterators. This version gathers all of the `anext` together to be ran concurrently. The downside is that it runs more tasks on the final iteration instead of stopping on the first one that finishes.